feat: 剧集序号支持小数 (#645)
* optimize: 剧集序号修改为浮点数

* build: gen new api-client v0.15.2+1

* optimize: console剧集编辑和新增弹框适应小数序号

* docs: update CHANGELOG.MD

+++ api/src/main/java/run/ikaros/api/core/subject/Episode.java
@@ -25,7 +25,7 @@ public class Episode {
     private String description;
     @JsonProperty("air_time")
     private LocalDateTime airTime;
-    private Integer sequence;
+    private Float sequence;
     private EpisodeGroup group;
 
     private List<EpisodeResource> resources;
@@ -36,7 +36,7 @@ public class Episode {
     public static Episode defaultEpisode(Long subjectId) {
         Episode episode = new Episode();
         episode.setAirTime(LocalDateTime.now());
-        episode.setSequence(1);
+        episode.setSequence(1F);
         episode.setGroup(EpisodeGroup.MAIN);
         episode.setDescription("Default episode description");
         episode.setName("Default episode name");
