+++ server/src/main/java/run/ikaros/server/store/entity/EpisodeEntity.java
@@ -27,7 +27,7 @@ public class EpisodeEntity extends BaseEntity {
     private String description;
     @Column("air_time")
     private LocalDateTime airTime;
-    private Integer sequence;
+    private Float sequence;
     @Column("ep_group")
     private EpisodeGroup group;
 }
